+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Change log for the little-timmy python module.
 ## [3.3.0] - 2025/06/27
 
 - Make template and playbook globs configurable
+- Look for templates in hidden files
 
 ## [3.2.0] - 2025/06/24